Extensions used in multiple places (potentially) moved to Extensions.swift (todo: document)
