+++ dialog/UI/FullscreenView.swift
@@ -9,18 +9,6 @@ import Foundation
 import SwiftUI
 import MarkdownUI
 
-extension Color {
-    init(hex: UInt, alpha: Double = 1) {
-        self.init(
-            .sRGB,
-            red: Double((hex >> 16) & 0xff) / 255,
-            green: Double((hex >> 08) & 0xff) / 255,
-            blue: Double((hex >> 00) & 0xff) / 255,
-            opacity: alpha
-        )
-    }
-}
-
 struct FullscreenView: View {
             
     @ObservedObject var observedData : DialogUpdatableContent
